+++ xdef/changelog.md
@@ -1,7 +1,10 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 42.1.4, release-date 2024-03-09
-* .
+# Version 42.1.4, release-date 2024-03-10
+* In the conversion of X-definition to XML schema (`org.xdef.util.XdefToXsd`)
+  is added parameter `--xx`. If this parameter is used than from xdatetime
+  X-definition validation method is used the mask from parameter `outFormat`
+ (the second sequential parameter) for validation of XML data.
 
 # Version 42.1.3, release-date 2024-03-04
 * The conversion of X-definition to XML schema (`org.xdef.util.XdefToXsd`)